Add publisher for target velocities
