+++ ros/src/computing/planning/decision/packages/decision_maker/CMakeLists.txt
@@ -39,6 +39,7 @@ include_directories(
 
 set(TARGET_SRC  nodes/decision_maker/decision_maker_node.cpp
 		nodes/decision_maker/decision_maker_node_callback.cpp
+		nodes/decision_maker/decision_maker_publish.cpp
 		nodes/decision_maker/cross_road_area.cpp
 	)
